rename: 아이콘 파일명 '-icon'로 통일
